# pkg/api/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
...
            "engine.SnapshotParams": {
                "properties": {
                    "include_ram": {
                        "type": "boolean"
                    },
                    "tag": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "jobs.Job": {
                "properties": {
                    "action": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    },
                    "finished_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    },
                    "logs": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "result": {
                        "type": "string"
                    },
                    "started_at": {
                        "type": "string"
                    },
                    "state": {
                        "$ref": "#/components/schemas/jobs.State"
                    },
                    "target": {
                        "type": "string"
                    }
                },
                "required": [
                    "action",
                    "created_at",
                    "id",
                    "label",
                    "logs",
                    "state",
                    "target"
                ],
                "type": "object"
            },
...
            "jobs.State": {
                "enum": [
                    "pending",
                    "running",
                    "succeeded",
                    "failed"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "Pending",
                    "Running",
                    "Succeeded",
                    "Failed"
                ]
            },
...
            "vm.Snapshot": {
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "has_ram": {
                        "type": "boolean"
                    },
                    "size_bytes": {
                        "type": "integer"
                    },
                    "tag": {
                        "type": "string"
                    }
                },
                "required": [
                    "created_at",
                    "has_ram",
                    "size_bytes",
                    "tag"
                ],
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer JWT obtained from POST /api/login.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "VM lifecycle, networking, media and jobs API. Mutations returning 202 are asynchronous; inspect the returned job for completion.",
        "title": "Maco API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/vms/{id}/snapshots": {
            "get": {
                "operationId": "listSnapshots",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/vm.Snapshot"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "listSnapshots",
                "tags": [
                    "vms"
                ]
            },
            "post": {
                "operationId": "createSnapshot",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/engine.SnapshotParams",
                                        "summary": "body",
                                        "description": "Request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "createSnapshot",
                "tags": [
                    "vms"
                ]
            }
        },
        "/api/vms/{id}/snapshots/{tag}": {
            "delete": {
                "operationId": "deleteSnapshot",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "tag",
                        "in": "path",
                        "name": "tag",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "deleteSnapshot",
                "tags": [
                    "vms"
                ]
            }
        },
        "/api/vms/{id}/snapshots/{tag}/restore": {
            "post": {
                "operationId": "restoreSnapshot",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "tag",
                        "in": "path",
                        "name": "tag",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jobs.Job"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "restoreSnapshot",
                "tags": [
                    "vms"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ]
}